refactor: 중복된 코드 제거
