+++ jdbc/src/main/java/org/springframework/jdbc/core/JdbcTemplate.java
@@ -1,10 +1,6 @@
 package org.springframework.jdbc.core;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
 import javax.sql.DataSource;
-import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -13,50 +9,24 @@ import java.util.List;
 
 public class JdbcTemplate {
 
-    private static final Logger log = LoggerFactory.getLogger(JdbcTemplate.class);
-
-    private final DataSource dataSource;
+    private final PreparedStatementExecutor preparedStatementExecutor;
 
     public JdbcTemplate(final DataSource dataSource) {
-        this.dataSource = dataSource;
+        this.preparedStatementExecutor = new PreparedStatementExecutor(dataSource);
     }
 
     public int update(final String sql, final Object... args) {
-        try (
-                final Connection conn = dataSource.getConnection();
-                final PreparedStatement pstmt = conn.prepareStatement(sql)
-        ) {
-            setArguments(pstmt, args);
-
-            log.debug("query : {}", sql);
-
-            return pstmt.executeUpdate();
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
-    }
-
-    private void setArguments(final PreparedStatement pstmt, final Object[] args) throws SQLException {
-        for (int i = 0; i < args.length; i++) {
-            final int parameterIndex = i + 1;
-            pstmt.setObject(parameterIndex, args[i]);
-        }
+        return preparedStatementExecutor.execute(PreparedStatement::executeUpdate, sql, args);
     }
 
     public <T> T queryForObject(final String sql, final RowMapper<T> rowMapper, final Object... args) {
-        try (
-                final Connection conn = dataSource.getConnection();
-                final PreparedStatement pstmt = conn.prepareStatement(sql)
-        ) {
-            setArguments(pstmt, args);
-            ResultSet rs = pstmt.executeQuery();
+        return preparedStatementExecutor.execute(
+                pstmt -> {
+                    final ResultSet rs = pstmt.executeQuery();
 
-            log.debug("query : {}", sql);
-
-            return calculateResult(rowMapper, rs);
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
+                    return calculateResult(rowMapper, rs);
+                }, sql, args
+        );
     }
 
     private <T> T calculateResult(final RowMapper<T> rowMapper, final ResultSet rs) throws SQLException {
@@ -68,22 +38,17 @@ public class JdbcTemplate {
     }
 
     public <T> List<T> query(final String sql, final RowMapper<T> rowMapper) {
-        try (
-                final Connection conn = dataSource.getConnection();
-                final PreparedStatement pstmt = conn.prepareStatement(sql)
-        ) {
-            ResultSet rs = pstmt.executeQuery();
+        return preparedStatementExecutor.execute(
+                pstmt -> {
+                    final ResultSet rs = pstmt.executeQuery();
 
-            log.debug("query : {}", sql);
-
-            return getResults(rowMapper, rs);
-        } catch (SQLException e) {
-            throw new RuntimeException(e);
-        }
+                    return getResults(rowMapper, rs);
+                }, sql
+        );
     }
 
     private <T> List<T> getResults(final RowMapper<T> rowMapper, final ResultSet rs) throws SQLException {
-        List<T> results = new ArrayList<>();
+        final List<T> results = new ArrayList<>();
 
         while (rs.next()) {
             final T result = calculateResult(rowMapper, rs);
