+++ jdbc/src/main/java/org/springframework/jdbc/core/PreparedStatementExecuteStrategy.java
@@ -1,0 +1,10 @@
+package org.springframework.jdbc.core;
+
+import java.sql.PreparedStatement;
+import java.sql.SQLException;
+
+@FunctionalInterface
+public interface PreparedStatementExecuteStrategy<T> {
+
+    T strategy(final PreparedStatement pstm) throws SQLException;
+}
